Remove print from check_repo
